+++ src/criteria/BlogPostCriteria.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * @copyright  Copyright (c) Flipbox Digital Limited
+ * @license    https://flipboxfactory.com/software/hubspot/license
+ * @link       https://www.flipboxfactory.com/software/hubspot/
+ */
+
+namespace flipbox\craft\hubspot\criteria;
+
+/**
+ * @author Flipbox Factory <hello@flipboxfactory.com>
+ * @since 1.0.0
+ */
+class BlogPostCriteria extends \Flipbox\HubSpot\Criteria\BlogPostCriteria
+{
+    use CacheTrait,
+        ConnectionTrait;
+}